[UPD] 🚨 Force pgbadger to run as sudo
Blue Stingray uses `docker` to run the Odoo (`web`) and PostgreSQL
(`db`) containers separately.

The `db` container's volume is mounted on the `web` container so that it
can be read from, but it is not accessible to the `odoo` user. To bypass
this permissions problem, the `pgbadger` command is run as the `root`
user.

If your configuration does not have this permissions issue or if you
want to resolve the permissions problem by managing a `999` user group,
then you can simply process a `git revert` on this commit to run the
`pgbadger` command as the non-`root` user.